refactor: V2 Team Service를 구현 계층으로 우선 분리
- TODO: 일차원적인 분리로 로직 흐름에 맞게 연관 서비스를 활용하는 식으로 리팩터링 필요
